feat(setup): refactor boot and shutdown logic, introduce server lifecycle management
